- Include skin tones in index - Use Uint32Array for index - Add `emojify()` and `codify()` convenience methods

+++ src/raw.d.ts
@@ -1,4 +1,4 @@
 export const POINTS: Uint32Array
-export const KEY_TO_POINTS: Uint16Array
+export const KEY_TO_POINTS: Uint32Array
 export const POINTS_TO_KEY: Uint16Array
 export const KEYS: Uint8Array
